% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_gene_pvals}
\alias{example_gene_pvals}
\title{Example gene_pvals named vector from mouse experiment.}
\format{
A named vector with 23,420 items
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE90693}
`GSE90693_RawCountsData_TPR50_6months_AllRegions.txt.gz`
}
\usage{
example_gene_pvals
}
\description{
A named vector containing gene symbols as name and p-values as values. This
is from a mouse Alzheimer's disease model (GEO: GSE90693), specifically 6
months after treatment in Frontal Cortex brain region. In this dataset, we
expect to see the Microglia as the most enriched cell type.
}
\keyword{datasets}
