% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRcell.R
\name{LRcell}
\alias{LRcell}
\title{Wrapper function to run LRcell for preranked gene set.}
\usage{
LRcell(
  gene_pvals,
  sc_marker_genes = NULL,
  sig_cutoff = 0.05,
  species = c("mouse", "human"),
  region = NULL,
  method = c("LR", "LiR"),
  min_size = 1,
  plot = FALSE
)
}
\arguments{
\item{gene_pvals}{Named vector of gene-level pvalues from DEG analysis, i.e.
DESeq2, LIMMA}

\item{sc_marker_genes}{List of Cell-type specific marker genes derived from
single-cell RNA-seq. The name of the list is cell-type or cluster name, the
values are marker genes vectors or numeric named vectors. LRcell provides
marker genes list in different human/mouse brains, but users could use their
own marker gene list as input.
default: NULL}

\item{sig_cutoff}{Cutoff for input genes' pvalues, default: 0.05.}

\item{species}{Either `mouse` or `human`, default: mouse.}

\item{region}{Specific brain regions provided by LRcell. For mouse, LRcell
provides 9 brain regions: c("FC", "HC", "PC", "GP", "STR", "TH", "SN", "ENT",
"CB"). For human, LRcell provides c("pFC")}

\item{method}{Either `logistic regression` or `linear regression`. Logistic
regression equally treats cell-type specific marker genes, however, if
certain values could determine the importance of marker genes, linear
regression can be performed, default: LR.}

\item{min_size}{Minimal size of a marker gene set, will impact the balance of
labels}

\item{plot}{Whether to generate the plot.}
}
\value{
LRcell statistics.
}
\description{
This function wraps around \link[LRcell]{LRcell_analysis} in case of empty
inputs of the marker gene file and brain region.
}
\examples{
data(example_gene_pvals)
res <- LRcell(example_gene_pvals, region="FC")
}
