% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRcell.R
\name{LRcell_analysis}
\alias{LRcell_analysis}
\title{Find most enriched cell types in bulk DE genes by Logistic Regression}
\usage{
LRcell_analysis(
  gene_pvals,
  sc_marker_genes,
  method,
  min_size = 1,
  sig_cutoff = 0.05,
  provided_flag = FALSE
)
}
\arguments{
\item{provided_flag}{Whether users are using provided marker genes to run
LRcell analysis.}
}
\description{
This is a function which takes marker genes from single-cell RNA-seq as
reference to calculate the enrichment of certain cell types in bulk DEG
analysis. We assume that bulk DEG is derived from certain cell-type specific
pattern.
}
