% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRcell.R
\name{LRcell_gene_enriched_scores}
\alias{LRcell_gene_enriched_scores}
\title{Find most enriched cell types in bulk DE genes by Logistic Regression}
\usage{
LRcell_gene_enriched_scores(expr, annot, power=1)
}
\arguments{
\item{expr}{Expression dataframe with rows as genes and columns as cells.}

\item{annot}{Cell type annotation named vector with names as cell ids and
values as cell types.}

\item{parallel}{Whether to run it in parallel.}

\item{cores.n}{Number of cores to run parallel, i.e. 4.}
}
\value{
Enrichment dataframe with rows as genes and columns as cell types,
values are enrichment scores.
}
\description{
This is a function which takes marker genes from single-cell RNA-seq as
reference to calculate the enrichment of certain cell types in bulk DEG
analysis. We assume that bulk DEG is derived from certain cell-type specific
pattern.
}
