% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRcell.R
\name{LRcell_helper}
\alias{LRcell_helper}
\title{Helper function to get provided single-cell marker genes}
\usage{
LRcell_helper(gene_pvals, method, species, region, min_size, sig_cutoff)
}
\description{
@param species `Mouse` or `Human`
 @param region Provided brain regions
 @return Provided single-cell marker genes list
}
