% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mouse_celltypes}
\alias{mouse_celltypes}
\title{Mapping between subclusters and cell types in Mouse Brain}
\format{
A named vector with 565 subclusters:
\describe{
    Named vector with name as subclusters and values as cell types.
}
}
\source{
\url{http://dropviz.org/} under tab `data`
}
\usage{
mouse_celltypes
}
\description{
A named vector containing the subclusters as name and cell types as values
in Mouse Brain. The cell types are pre-annotated by the dataset, which
includes: Endothelial, FibroblastLike, Mural, Oligodendrocytes,
Polydendrocytes, Astrocytes and Microglia.
}
\keyword{datasets}
