% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mouse_FC_marker_genes}
\alias{mouse_FC_marker_genes}
\title{Example marker genes from mouse FC brain region.}
\format{
A named vector with 81 subclusters in mouse Frontal Cortex:
\describe{
    Named vector with name as subclusters and values as marker genes.
}
}
\source{
Caculated from gene enrichment scores
}
\usage{
data(mouse_FC_marker_genes)
}
\description{
A list of marker genes with names indicating cell types. We selected top 100
enriched genes from each subcluster as marker genes list.
}
\keyword{datasets}
