% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{MOUSE_REGIONS}
\alias{MOUSE_REGIONS}
\title{mouse brain regions}
\format{
An object of class \code{character} of length 9.
}
\usage{
MOUSE_REGIONS
}
\description{
mouse brain regions
}
\keyword{datasets}
