% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_gsea_enrich}
\alias{plot_gsea_enrich}
\title{GSEA enrichment plot for the enrichment of different cell types}
\usage{
plot_gsea_enrich(gene_pvals, sc_marker_genes)
}
\description{
GSEA enrichment plot for the enrichment of different cell types
}
