% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rename_by_extdata}
\alias{rename_by_extdata}
\title{Change the cluster number into corresponding sub-cell types}
\usage{
rename_by_extdata()
}
\description{
Change the cluster number into corresponding sub-cell types
}
