% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rename_clusters}
\alias{rename_clusters}
\title{Curate cluster name according to mouse cell types}
\usage{
rename_clusters(enriched_genes)
}
\arguments{
\item{enriched_genes}{The provided enriched genes.}
}
\description{
Curate cluster name according to mouse cell types
}
