% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_region}
\alias{validate_region}
\title{Whether the input region is valid}
\usage{
validate_region(species, region)
}
\arguments{
\item{species}{Input species.}

\item{region}{Input region.}
}
\description{
Whether the input region is valid
}
