% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_markergenes}
\alias{get_markergenes}
\title{Get top marker genes for each subcluster}
\usage{
get_markergenes(enriched_genes, method = c("LR", "LiR"), topn = 100)
}
\description{
Get top marker genes for each subcluster
}
