% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRcell.R
\name{LRcell}
\alias{LRcell}
\title{Find transcriptomicly active cell types by Logistic Regression}
\usage{
LRcell(
  gene_pvals,
  sc_marker_genes = NULL,
  species = c("mouse", "human"),
  region = NULL,
  method = c("LR", "LiR"),
  plot = FALSE
)
}
\arguments{
\item{gene_pvals}{Named vector of gene-level pvalues from DEG analysis, i.e.
DESeq2, LIMMA}

\item{sc_marker_genes}{List of Cell-type specific marker genes derived from
single-cell RNA-seq. The name of the list is cell-type or cluster name, the
values are marker genes vectors or numeric named vectors. LRcell provides
marker genes list in different human/mouse brains, but users could use their
own marker gene list as input.
default: NULL}

\item{species}{Either `mouse` or `human`, default: mouse}

\item{region}{Specific brain regions provided by LRcell. For mouse, LRcell
provides 9 brain regions according to Saunders et al. (Cell 2018) including:
Frontal Cortex, Posterior Cortex, Hippocampus, etc. For human, LRcell
provides pre-frontal cortex (Nagy et al., Nature Neuroscience 2020)}

\item{method}{Either `logistic regression` or `linear regression`. Logistic
regression equally treats cell-type specific marker genes, however, if
certain values could determine the importance of marker genes, linear
regression can be performed.}

\item{plot}{For Manhattan and Enrichment plot generation}
}
\value{
LRcell statistics.
}
\description{
This is a function which takes marker genes from single-cell RNA-seq as
reference to calculate the enrichment of certain cell types in bulk DEG
analysis. We assume that bulk DEG is derived from certain cell-type specific
pattern.
}
