% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_gsea_enrich}
\alias{plot_gsea_enrich}
\title{GSEA enrichment plot for the enrichment of different cell types}
\usage{
plot_gsea_enrich(gene_pvals, sc_marker_genes)
}
\arguments{
\item{gene.p}{Gene pvalues}

\item{marker.g}{}
}
\value{
A ggplot2 object
}
\description{
This function draws out the LRcell result dataframe. In this function, we take
LRcell result dataframe and added cell types according to
}
