% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_manhattan_enrich}
\alias{plot_manhattan_enrich}
\title{Manhattan plot for the enrichment of cell types}
\usage{
plot_manhattan_enrich(lrcell_res)
}
\arguments{
\item{lrcell_res}{LRcell result dataframe}
}
\value{
A ggplot2 object
}
\description{
This function draws out the LRcell result dataframe. In this function, we take
LRcell result dataframe and added cell types according to
}
